more organization of App file, add separate AppDrawer
