+++ packages/unigraph-dev-explorer/src/components/AppDrawer.tsx
@@ -1,0 +1,54 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+import Divider from '@material-ui/core/Divider';
+import Drawer from '@material-ui/core/Drawer';
+import List from '@material-ui/core/List';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemIcon from '@material-ui/core/ListItemIcon';
+import ListItemText from '@material-ui/core/ListItemText';
+import { makeStyles } from '@material-ui/core/styles';
+import { Home, Info } from '@material-ui/icons';
+
+const drawerWidth = 240;
+
+const useStyles = makeStyles(theme => ({
+  drawer: {
+    width: drawerWidth,
+    flexShrink: 0,
+  },
+  drawerPaper: {
+    width: drawerWidth,
+  },
+  toolbar: {
+    minHeight: '48px !important'
+  },
+}));
+
+export default function DrawerRouter() {
+  const classes = useStyles();
+
+  return (
+    <Drawer
+      className={classes.drawer}
+      variant="permanent"
+      classes={{
+        paper: classes.drawerPaper,
+      }}
+      anchor="left"
+    >
+      <div className={classes.toolbar} />
+      <Divider />
+      <List>
+        <ListItem button component={Link} to="/">
+          <ListItemIcon><Home /></ListItemIcon>
+          <ListItemText primary="Home" />
+        </ListItem>
+        <ListItem button component={Link} to="/about">
+          <ListItemIcon><Info /></ListItemIcon>
+          <ListItemText primary="About" />
+        </ListItem>
+      </List>
+    </Drawer>
+  );
+}
